requirements.txt
pandas~=2.2.2
seaborn~=0.13.2
matplotlib~=3.9.1
numpy~=1.26.4
scikit-learn~=1.5.1
xgboost~=2.1.0
yagmail~=0.15.293
uvicorn~=0.30.3
fastapi~=0.111.1
APScheduler~=3.10.4
python-dotenv~=1.0.1

pydantic~=2.8.2